Extract plots from Chen2021FitnessResults
