+++ src/models/chen2021Fitness/Chen2021FitnessContainer.tsx
@@ -8,11 +8,6 @@ import { Chen2021FitnessResults } from './Chen2021FitnessResults';
 
 type ContainerProps = zod.infer<typeof SampleSelectorSchema>;
 
-const Wrapper = styled.div`
-  background-color: #ffe0b6;
-  padding: 15px;
-`;
-
 const SectionHeader = styled.h5`
   margin-top: 20px;
 `;
@@ -68,8 +63,7 @@ export const Chen2021FitnessContainer = ({
   };
 
   return (
-    <Wrapper>
-      <h4>Fitness Advantage Estimation</h4>
+    <>
       <p>
         The model assumes that the increase or decrease of the proportion of a variant follows a logistic
         function. It fits a logistic model to the data by optimizing the maximum likelihood to obtain the
@@ -130,6 +124,6 @@ export const Chen2021FitnessContainer = ({
           </a>
         </li>
       </ul>
-    </Wrapper>
+    </>
   );
 };
